Feature/1135 formatter enable rules rebased (#1260)
* apply eslint

squash branch

update eslint config

run formatter

fix tests

revert to single quotes

regen package-lock

move eslint dependendencies to devDependencies and update comment in eslint-config

remove not needed parent clings

fix typo in comment in scoring component tests

put both statements inside if statement

change asignment of args in global ts

add curly to eslint config

put curly into quotes

update comments to clarify that they are for lint

replace multiple ifs with one switch

replace multiple ifs with one switch

replace multiple ifs with one switch

turn no-non-null-assertion off and bring back all usages of

add strict equality

fix eslint errors

remove not needed comment

add todo for error handling oauth interceptor

update comments so its clear that linter needs them

revert change made accidentally

fix frontend unit tests

revert changes made accidentally

bring back all not null assertions and create some switch statements in scoring component for readability

* fix formatting errors

* regen package-lock

---------

+++ frontend/src/app/team-management/member-detail/member-detail.component.ts
@@ -48,11 +48,10 @@ export class MemberDetailComponent implements OnInit, OnDestroy {
 
   ngOnInit(): void {
     this.route.paramMap
-      .pipe(takeUntil(this.unsubscribe$),
-        tap((params) => {
-          const id = this.getIdFromParams(params);
-          this.loadUser(id);
-        }))
+      .pipe(takeUntil(this.unsubscribe$), tap((params) => {
+        const id = this.getIdFromParams(params);
+        this.loadUser(id);
+      }))
       .subscribe();
   }
 
@@ -92,8 +91,7 @@ export class MemberDetailComponent implements OnInit, OnDestroy {
     this.dialogService
       .openConfirmDialog('CONFIRMATION.DELETE.USER_FROM_TEAM', i18nData)
       .afterClosed()
-      .pipe(filter((confirm) => confirm),
-        mergeMap(() => this.teamService.removeUserFromTeam(user.id, userTeam.team)))
+      .pipe(filter((confirm) => confirm), mergeMap(() => this.teamService.removeUserFromTeam(user.id, userTeam.team)))
       .subscribe(() => {
         this.loadUser(user.id);
         this.userService.reloadUsers();
